emotion-dubbing/docker/emotion-features/scripts/parallel_inference.py
import asyncio
import json
from typing import List, Dict
import torch
from concurrent.futures import ThreadPoolExecutor

class ModelRunner:
    def __init__(self, model_configs: List[Dict]):
        self.models = {}
        self.executor = ThreadPoolExecutor()
        for config in model_configs:
            self.load_model(config)
    
    def load_model(self, config: Dict):
        model_type = config['type']
        if model_type == 'whisper':
            import whisper
            self.models[model_type] = whisper.load_model(config['name'])
        elif model_type == 'wav2vec2':
            from transformers import Wav2Vec2ForSequenceClassification
            self.models[model_type] = Wav2Vec2ForSequenceClassification.from_pretrained(config['name'])

    async def run_parallel(self, audio_path: str):
        tasks = []
        for model_name, model in self.models.items():
            tasks.append(self.executor.submit(self.process_audio, model_name, model, audio_path))
        
        results = {}
        for task in tasks:
            result = await asyncio.wrap_future(task)
            results.update(result)
        
        return results

    def process_audio(self, model_name: str, model: torch.nn.Module, audio_path: str):
        # Implementation for each model type
        pass